增加了会员操作日志、会员动态 解决了会员在会员中心修改密码双重加密问题

+++ src/main/java/com/lxinet/jeesns/modules/cms/service/impl/ArticleCommentServiceImpl.java
@@ -3,6 +3,7 @@ package com.lxinet.jeesns.modules.cms.service.impl;
 import com.lxinet.jeesns.core.dto.ResponseModel;
 import com.lxinet.jeesns.core.entity.Page;
 import com.lxinet.jeesns.core.interceptor.PageInterceptor;
+import com.lxinet.jeesns.core.utils.ActionUtil;
 import com.lxinet.jeesns.core.utils.StringUtils;
 import com.lxinet.jeesns.modules.cms.dao.IArticleCommentDao;
 import com.lxinet.jeesns.modules.cms.dao.IArticleDao;
@@ -11,6 +12,7 @@ import com.lxinet.jeesns.modules.cms.entity.ArticleComment;
 import com.lxinet.jeesns.modules.cms.service.IArticleCommentService;
 import com.lxinet.jeesns.modules.cms.service.IArticleService;
 import com.lxinet.jeesns.modules.mem.entity.Member;
+import com.lxinet.jeesns.modules.sys.service.IActionLogService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,6 +29,8 @@ public class ArticleCommentServiceImpl implements IArticleCommentService {
     private IArticleCommentDao articleCommentDao;
     @Resource
     private IArticleService articleService;
+    @Resource
+    private IActionLogService actionLogService;
 
     @Override
     public ArticleComment findById(int id) {
@@ -70,9 +74,14 @@ public class ArticleCommentServiceImpl implements IArticleCommentService {
 
     @Override
     @Transactional
-    public ResponseModel delete(int id) {
+    public ResponseModel delete(Member loginMember, int id) {
+        ArticleComment articleComment = this.findById(id);
+        if(articleComment == null){
+            return new ResponseModel(-1,"评论不存在");
+        }
         int result = articleCommentDao.delete(id);
         if(result == 1){
+            actionLogService.save(loginMember.getCurrLoginIp(),loginMember.getId(), ActionUtil.DELETE_ARTICLE_COMMENT,"ID："+articleComment.getId()+"，内容："+articleComment.getContent());
             return new ResponseModel(1,"删除成功");
         }
         return new ResponseModel(-1,"删除失败");
