+++ src/main/java/com/lxinet/jeesns/modules/group/service/impl/GroupTopicCommentServiceImpl.java
@@ -3,6 +3,7 @@ package com.lxinet.jeesns.modules.group.service.impl;
 import com.lxinet.jeesns.core.dto.ResponseModel;
 import com.lxinet.jeesns.core.entity.Page;
 import com.lxinet.jeesns.core.interceptor.PageInterceptor;
+import com.lxinet.jeesns.core.utils.ActionUtil;
 import com.lxinet.jeesns.core.utils.StringUtils;
 import com.lxinet.jeesns.modules.group.dao.IGroupTopicCommentDao;
 import com.lxinet.jeesns.modules.group.entity.GroupTopic;
@@ -10,6 +11,7 @@ import com.lxinet.jeesns.modules.group.entity.GroupTopicComment;
 import com.lxinet.jeesns.modules.group.service.IGroupTopicCommentService;
 import com.lxinet.jeesns.modules.group.service.IGroupTopicService;
 import com.lxinet.jeesns.modules.mem.entity.Member;
+import com.lxinet.jeesns.modules.sys.service.IActionLogService;
 import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import java.util.List;
@@ -23,6 +25,8 @@ public class GroupTopicCommentServiceImpl implements IGroupTopicCommentService {
     private IGroupTopicCommentDao groupTopicCommentDao;
     @Resource
     private IGroupTopicService groupTopicService;
+    @Resource
+    private IActionLogService actionLogService;
 
     @Override
     public GroupTopicComment findById(int id) {
@@ -65,9 +69,14 @@ public class GroupTopicCommentServiceImpl implements IGroupTopicCommentService {
     }
 
     @Override
-    public ResponseModel delete(int id) {
+    public ResponseModel delete(Member loginMember,int id){
+        GroupTopicComment groupTopicComment = this.findById(id);
+        if(groupTopicComment == null){
+            return new ResponseModel(-1,"评论不存在");
+        }
         int result = groupTopicCommentDao.delete(id);
         if(result == 1){
+            actionLogService.save(loginMember.getCurrLoginIp(),loginMember.getId(), ActionUtil.DELETE_GROUP_TOPIC_COMMENT,"ID："+groupTopicComment.getId()+"，内容："+groupTopicComment.getContent());
             return new ResponseModel(1,"删除成功");
         }
         return new ResponseModel(-1,"删除失败");
